rejig project structure - moved components into /components
