+++ src/explore-education-statistics-admin/src/components/footnotes/FootnoteForm.module.scss
