+++ src/explore-education-statistics-admin/src/components/footnotes/FootnoteSubjectSelection.tsx
