+++ src/explore-education-statistics-admin/src/components/footnotes/FootnotesList.module.scss
