+++ src/explore-education-statistics-admin/src/components/footnotes/FootnotesList.tsx
